Create stickTest.html
*physics is commented out so that you can see .pos is working - conflict w .loc(x,y) i believe...

More obvious when shapes are rectangles because you can reveal the red shape, but the join is functioning.

I did a lot of copy pasting from "chimes.html" and "physics.html" sessions from class. SO there may be some redundancies but i think it's good.

To still be solved: located them next to each other rather than on top.
Then to apply on our semi class - hence the importance of placing correctly

Sam & Ralph are the sheepdog and wolf from looney tunes